requirements.txt
asgiref==3.7.2
certifi==2025.4.26
charset-normalizer==3.4.2
dj-database-url==2.2.0
Django==3.2.25
django-ckeditor==6.3.2
django-js-asset==2.0.0
gunicorn==23.0.0
idna==3.10
importlib-metadata==6.7.0
packaging==24.0
Pillow==9.5.0
psycopg2==2.9.9
python-decouple==3.8
pytz==2025.2
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.7
whitenoise==6.5.0
zipp==3.15.0
